Update with spectral similarity
